+++ tools/extensions/kimapp-snippets/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to the "kimapp-snippets" extension will be documented in this file.
 
-## [1.0.0] - 2024-07-23
+## [1.0.0]
 
-- Initial release
-- Added Autoverpod snippet for creating form widget classes 
+- Initial release with code snippets for Kimapp development
+
+## [1.1.0]
+
+- Added feature generation functionality
+- Right-click on a directory in the file explorer to create a new feature
+- Generates all necessary files and boilerplate code for a new feature